+++ core/src/note.rs
@@ -14,7 +14,7 @@ use dusk_bytes::{DeserializableSlice, Error as BytesError, Serializable};
 use dusk_jubjub::{dhke, JubJubAffine, JubJubScalar, GENERATOR_NUMS_EXTENDED};
 use dusk_poseidon::{Domain, Hash};
 use ff::Field;
-use jubjub_elgamal as elgamal;
+use jubjub_elgamal::{DecryptFrom, Encryption as ElGamal};
 use jubjub_schnorr::{PublicKey as NotePublicKey, SecretKey as NoteSecretKey};
 use rand::{CryptoRng, RngCore};
 
@@ -476,16 +476,24 @@ impl Sender {
         sender_pk: &PublicKey,
         blinder: &[JubJubScalar; 2],
     ) -> Self {
-        let sender_enc_A =
-            elgamal::encrypt(note_pk.as_ref(), sender_pk.A(), &blinder[0]);
+        let (sender_enc_A, _) = ElGamal::encrypt(
+            note_pk.as_ref(),
+            sender_pk.A(),
+            None,
+            &blinder[0],
+        );
 
-        let sender_enc_B =
-            elgamal::encrypt(note_pk.as_ref(), sender_pk.B(), &blinder[1]);
+        let (sender_enc_B, _) = ElGamal::encrypt(
+            note_pk.as_ref(),
+            sender_pk.B(),
+            None,
+            &blinder[1],
+        );
 
         let sender_enc_A: (JubJubAffine, JubJubAffine) =
-            (sender_enc_A.0.into(), sender_enc_A.1.into());
+            (sender_enc_A.c1().into(), sender_enc_A.c2().into());
         let sender_enc_B: (JubJubAffine, JubJubAffine) =
-            (sender_enc_B.0.into(), sender_enc_B.1.into());
+            (sender_enc_B.c1().into(), sender_enc_B.c2().into());
 
         Self::Encryption([sender_enc_A, sender_enc_B])
     }
@@ -507,17 +515,15 @@ impl Sender {
             }
         };
 
-        let sender_enc_A = sender_enc[0];
-        let sender_enc_B = sender_enc[1];
+        let sender_enc_A =
+            ElGamal::new(sender_enc[0].0.into(), sender_enc[0].1.into());
+        let sender_enc_B =
+            ElGamal::new(sender_enc[1].0.into(), sender_enc[1].1.into());
 
-        let decrypt_A = elgamal::decrypt(
-            note_sk.as_ref(),
-            &(sender_enc_A.0.into(), sender_enc_A.1.into()),
-        );
-        let decrypt_B = elgamal::decrypt(
-            note_sk.as_ref(),
-            &(sender_enc_B.0.into(), sender_enc_B.1.into()),
-        );
+        let decrypt_A =
+            sender_enc_A.decrypt(&DecryptFrom::SecretKey(*note_sk.as_ref()));
+        let decrypt_B =
+            sender_enc_B.decrypt(&DecryptFrom::SecretKey(*note_sk.as_ref()));
 
         Ok(PublicKey::new(decrypt_A, decrypt_B))
     }